Add files via upload
didn't think i'd ever update this again! it occurred to me that i should add distrito, zona, (precinct) latitude, and (precinct) longitude to the 'final_computo' data file. they were on geoelectoral, but not in the election data. also, i changed html file to include distrito and zona as geographic levels, but i'd have to add some extra logic to deal with ones with no district/no zone. these are given as '', which will get mistreated as a real place. the right thing would be to match within the entire locality for those. don't feel like writing that, which would require me to enumerate all the leaves, since none of this matters now and i have much better prediction algos elsewhere.